Implement `NGon`s using constant generics.
This change upgrades the dependency on the `arrayvec` crate and
implements `NGon`s using constant generics instead of an opaque type
implementing the `Array` trait. Additionally, extended `Iterator` APIs
are used instead of `theon` traits for de/composition of arrays and
tuples to and from iterators.

Note that the minimum support Rust version is bumped to 1.51.0 for this
change.