Implement `NGon` and `Flat` with constant generics.
This change introduces constant generics and uses them to implement the
`NGon` and `Flat` APIs. As part of this change, the dependency on the
`arrayvec` crate has been upgraded and the `Array` trait is no longer
used. The minimum supported Rust version has been bumped to 1.51.0 for
these changes.

Because constraints and computations on constant generics are not yet
supported, traits for morphisms between `typenum` types and `usize`
constant have also be introduced in the `constant` module. This allows
for relations and computations on constants in APIs.

Additionally, this change improves some documentation and prefers a new
`TryFromIterator` trait over the dubious `FromItems` and `IntoItems`
traits from the `theon` crate.