Fix and improve mutable circulator test.
This change constructs a simple graph and uses circulators to write and
then read and verify data in the graph. Previously, this test did not
actually exercise circulators and the unsafe code local to the crate,
because it used a categorical iterator that mapped the items of an
external iterator type (not circulators).