Remove unstable GATs implementation.
This change migrates to a stable GATs implementation for entity storage
(i.e., graphs) and removes the `unstable` Cargo feature and its
corresponding `build.rs`.

As part of this change, the input type parameter `G` used in many view
implementations has been removed anywhere the compiler required a
lifetime bound. The `GraphData` type parameter is not actually composed
into view types, but the compiler cannot determine this. As such, any
bounds on this type should be avoided. The parameter `G` may be
completely removed in a subsequent change, because while it sometimes
makes the code a bit easier to read, it may also cause confusion.